chore: automatic npm install setup workaround
The project failed to setup correctly on the first try. The ideal
solution would be for the project image to contain the deps or build
them during the build step, but this is a quick solution that doesn't
require deep understanding of the node image.

To use this, I added an entry to my Reaction platform's
`config.local.mk`:

```
define SUBPROJECT_REPOS
\# other projects...
git@github.com:/reactioncommerce/federated-gateway.git,federated-gateway,trunk
endef
```

And then I can start the project with platform. All setup and
`npm installing` runs automatically.

```
make init-federated-gateway
```

Signed-off-by: Ticean Bennett <ticean@reactioncommerce.com>